Changed the way workspaces and, buttons are stored
Workspaces in i3wm-delegate are now stored in a GSList sorted by the
workspace's num and name attributes. In i3w-plugin the button list was
replaced by a GHashTable which maps the workspace to the button
associated with it.
Fix a bunch of coding style issues.
On urgent event the urgent handler is now only called for the
appropriate workspaces.

Scrolling crashes the plugin, also the urgent, rename, init and empty
events were not tested yet.