Added multi-monitor-utils sources and added it to Makefile.am
The multi-monitor-utils file defines two functions which allow for querying
the X Server for monitor information (output names and position/resolution).
And performing bounds-check to see in which output a window (by coordinates)
is located. This information should be the same that i3 is using, since it
gets it from the same place.